Refactor LLVM Python libraries to prepare for JVM
